+++ assets/javascripts/kitten/components/molecules/buttons/button/index.js
@@ -426,6 +426,7 @@ Button.propTypes = {
     'boron',
     'neon',
     'iron',
+    'calcium',
     'social_facebook',
     'social_twitter',
     'social_linkedin',