+++ src/esm/components/molecules/buttons/button/helpers/modifier-styles.js
@@ -100,6 +100,18 @@ export var modifierStyles = function modifierStyles(modifier) {
       activeColor = COLORS.background1;
       break;
 
+    case 'calcium':
+      borderColor = COLORS.line1;
+      backgroundColor = COLORS.background1;
+      color = COLORS.font1;
+      hoverBorderColor = COLORS.primary4;
+      hoverBgColor = COLORS.background1;
+      hoverColor = COLORS.font1;
+      activeBorderColor = COLORS.primary1;
+      activeBgColor = COLORS.primary5;
+      activeColor = COLORS.font1;
+      break;
+
     /* Social modifiers */
 
     case 'social_facebook':